web: support arrow keys for movement
Since findDOMNode is removed in React 19, I had to use an array of DOM refs.
This seems a *little* wrong but it works since I always generate the same
number of children.

It also occurs to me that I could have skipped all of this if I'd just used
tabindex as it's intended instead of forcing it to be mouse-only...