board: prevent panic when "solving" incorrectly
If the cell was not already solved and also did not have a particular bit set,
then it ends up filtering-out the same bit in the conflicting cell.  This
leaves `allowed` with the value 0x1, which panics when it masks and
`.count_ones()`.

Instead, just throw an exception to JS, which at least leaves the Wasm
interpreter functional.