web: fix a couple oddities with mouseover
Turns out that giving React a key to work with is indeed important ... since
there's no natural identity of the results of rules, for now Rust can just spit
out an entire serialization of the hint that will remain unique.

And for reasons that I don't understand, mouseover *does* happen when the DOM
shifts around when a hint is deleted (i.e. when it's applied and no longer
relevant), whereas mouseenter seems not to fire.  This is confusing to me
because mouseenter is documented to behave more sensibly...