rules: deduplicate DigitsCovered hints
This was observed most often when a cell was narrowed down to one choice, and
that same value was allowed for a cell that is in the same row/column AND in
the same block.

I can't immediately figure out how to construct an example for getting an
exactly-matching set of affected cells with any other state, so perhaps I can
just not do this BTreeSet-keeping.  DigitsCovered should exclude causes that
are narrowed down to one choice, because they're better covered by
OnlyOneAllowedValue anyway.