web, board: replace undo stack with history API
The whole state of the board can be stored in a small-ish base64 blob and
passed freely back and forth to Javascript.  We can just push that state to the
history stack every time something changes.

Since we need to recreate the board via `from_slug()`, the board itself has to
move to a useState to trigger re-render.  As a result, `cells` can be more
cleanly modeled with `useSyncExternalStore` since that gives us a way to
trigger React when we *don't* need to replace the whole board object.